chore: bridge qb-inventory calls to ox
